requirements.txt
python-dotenv==1.0.1
pyinstaller==6.3.0
requests==2.31.0
